fix dtype of y as float64

+++ delta_data_preprocess.py
@@ -71,7 +71,7 @@ def delta_data_converter(x, y, tf_list, data_order):
                 update_progress_bar(i / esti_events_num * 100, '{}/{}'.format(i, esti_events_num))
     print('\nEvents processed: {} , DONE'.format(i))
     x = new_x[~np.isnan(new_x).any(axis=1)]
-    y = np.asarray(new_y, dtype='float16')
+    y = np.asarray(new_y, dtype='float64')
 
     return x, y
